-Centralise and improve super cache handling fix. -Improve NodeTypeEnums to contain more information (at no cost) and replace as many instanceof as possible.

+++ drools-base/src/main/java/org/drools/base/common/NetworkNode.java
@@ -21,7 +21,6 @@ package org.drools.base.common;
 import java.io.Serializable;
 
 import org.drools.base.reteoo.BaseTerminalNode;
-import org.drools.core.reteoo.ObjectTypeNodeId;
 import org.kie.api.definition.rule.Rule;
 
 /**
@@ -40,7 +39,7 @@ public interface NetworkNode extends Serializable {
      */
     RuleBasePartitionId getPartitionId();
     
-    short getType();
+    int getType();
 
     Rule[] getAssociatedRules();
