+++ drools-core/src/main/java/org/drools/core/common/QuadroupleNonIndexSkipBetaConstraints.java
@@ -69,11 +69,11 @@ public class QuadroupleNonIndexSkipBetaConstraints
         return this;
     }
 
-    public void init(BuildContext context, short betaNodeType) {
+    public void init(BuildContext context, int betaNodeType) {
         constraints.init(context, betaNodeType);
     }
 
-    public void initIndexes(int depth, short betaNodeType, RuleBaseConfiguration config) {
+    public void initIndexes(int depth, int betaNodeType, RuleBaseConfiguration config) {
         constraints.initIndexes(depth, betaNodeType, config);
     }
 
@@ -118,7 +118,7 @@ public class QuadroupleNonIndexSkipBetaConstraints
     }
 
     public BetaMemory createBetaMemory(final RuleBaseConfiguration config,
-                                       final short nodeType) {
+                                       final int nodeType) {
         return constraints.createBetaMemory(config,
                                             nodeType);
     }
