+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/MockBetaNode.java
@@ -30,7 +30,6 @@ import org.drools.core.reteoo.LeftTupleSource;
 import org.drools.core.reteoo.ModifyPreviousTuples;
 import org.drools.core.reteoo.ObjectSource;
 import org.drools.core.reteoo.ReteooBuilder;
-import org.drools.core.reteoo.RightTuple;
 import org.drools.core.reteoo.RuleRemovalContext;
 import org.drools.core.reteoo.Sink;
 import org.drools.core.reteoo.TupleImpl;
@@ -83,7 +82,7 @@ public class MockBetaNode extends BetaNode {
                                   final ReteEvaluator reteEvaluator) {
     }
 
-    public short getType() {
+    public int getType() {
         return 0;
     }
